Measurements with the same units are conisdered identical types of measurements, even if the template parameters are in different orders

+++ include/MetaMeasure/UnitBase.hpp
@@ -4,7 +4,7 @@
 namespace MetaMeasure
 {
 
-using ExponentType = short;
+using ExponentType = int;
 
 template<typename IdentifierT, ExponentType ExponentV = 1>
 struct Dimension
